chore: migrate React components to TSX and enforce strict mode
