+++ eslint.config.js
@@ -1,2 +1,9 @@
-import config from "./.eslintrc.json" assert { type: "json" };
-export default config;
+import js from "@eslint/js";
+import react from "eslint-plugin-react";
+import reactHooks from "eslint-plugin-react-hooks";
+
+export default [
+  js.configs.recommended,
+  react.configs.recommended,
+  reactHooks.configs.recommended,
+];
