+++ src/components/LanguageSelector.tsx
@@ -1,11 +1,11 @@
 /**
  * Selector de idioma mejorado
  * Muestra y permite cambiar entre los idiomas disponibles
- * 
+ *
  * File: src/components/LanguageSelector.jsx
  */
-import React, { useState, useRef, useEffect } from 'react';
-import { useI18n } from '../utils/i18n';
+import React, { useState, useRef, useEffect } from "react";
+import { useI18n } from "../utils/i18n";
 
 const LanguageSelector = () => {
   const [isOpen, setIsOpen] = useState(false);
@@ -16,13 +16,15 @@ const LanguageSelector = () => {
 
   // Opciones de idioma con banderas y nombres
   const languageOptions = [
-    { code: 'en', name: 'English', flag: '🇬🇧' },
-    { code: 'es', name: 'Español', flag: '🇪🇸' },
-    { code: 'fr', name: 'Français', flag: '🇫🇷' }
+    { code: "en", name: "English", flag: "🇬🇧" },
+    { code: "es", name: "Español", flag: "🇪🇸" },
+    { code: "fr", name: "Français", flag: "🇫🇷" },
   ];
 
   // Encontrar el idioma actual
-  const currentLanguage = languageOptions.find(option => option.code === lang) || languageOptions[0];
+  const currentLanguage =
+    languageOptions.find((option) => option.code === lang) ||
+    languageOptions[0];
 
   // Cerrar al hacer clic fuera del dropdown
   useEffect(() => {
@@ -32,9 +34,9 @@ const LanguageSelector = () => {
       }
     };
 
-    document.addEventListener('mousedown', handleClickOutside);
+    document.addEventListener("mousedown", handleClickOutside);
     return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
+      document.removeEventListener("mousedown", handleClickOutside);
     };
   }, []);
 
@@ -49,28 +51,28 @@ const LanguageSelector = () => {
     let path = window.location.pathname;
 
     // Quitar prefijos de idioma existentes
-    const langPrefixes = ['/es/', '/fr/'];
+    const langPrefixes = ["/es/", "/fr/"];
     for (const prefix of langPrefixes) {
       if (path.startsWith(prefix)) {
         path = path.substring(prefix.length - 1); // Mantener la barra inicial
         break;
       }
     }
-    if (path === '/es' || path === '/fr') {
-      path = '/';
+    if (path === "/es" || path === "/fr") {
+      path = "/";
     }
 
     // Normalizar path
-    if (!path || path === '') path = '/';
+    if (!path || path === "") path = "/";
 
     // Construir nueva URL
     let newUrl;
-    if (langCode === 'en') {
+    if (langCode === "en") {
       // Para inglés (idioma por defecto) usamos la ruta sin prefijo
       newUrl = path;
     } else {
       // Para otros idiomas, añadir el prefijo de idioma
-      newUrl = `/${langCode}${path === '/' ? '' : path}`;
+      newUrl = `/${langCode}${path === "/" ? "" : path}`;
     }
 
     // Navegar a la nueva URL
@@ -89,14 +91,21 @@ const LanguageSelector = () => {
         aria-haspopup="listbox"
       >
         <span className="text-sm mr-1">{currentLanguage.flag}</span>
-        <span className="text-sm font-medium">{currentLanguage.code.toUpperCase()}</span>
+        <span className="text-sm font-medium">
+          {currentLanguage.code.toUpperCase()}
+        </span>
         <svg
-          className={`w-4 h-4 transition-transform ${isOpen ? 'rotate-180' : ''}`}
+          className={`w-4 h-4 transition-transform ${isOpen ? "rotate-180" : ""}`}
           fill="none"
           viewBox="0 0 24 24"
           stroke="currentColor"
         >
-          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M19 9l-7 7-7-7"
+          />
         </svg>
       </button>
 
@@ -113,7 +122,7 @@ const LanguageSelector = () => {
                 key={option.code}
                 onClick={() => handleLanguageChange(option.code)}
                 className={`w-full text-left px-4 py-2 text-sm flex items-center gap-2 
-                         ${lang === option.code ? 'bg-brand-red/10 text-brand-red' : 'text-gray-700 dark:text-gray-200'} 
+                         ${lang === option.code ? "bg-brand-red/10 text-brand-red" : "text-gray-700 dark:text-gray-200"} 
                          hover:bg-gray-100 dark:hover:bg-gray-700 transition-colors`}
                 role="option"
                 aria-selected={lang === option.code}
