+++ src/components/OptimizedProfileImage.tsx
