+++ src/components/StructuredData.tsx
