+++ tsconfig.json
@@ -11,19 +11,14 @@
     "src/assets/**/*.{ts,tsx,js,jsx,astro}",
     "src/env.d.ts"
   ],
-  "exclude": [
-    "dist",
-    "node_modules",
-    "src/components/ContactForm.jsx",
-    "src/components/Icon.jsx",
-    "src/components/Icons.jsx",
-    "**/*.spec.ts",
-    "**/*.test.ts"
-  ],
+  "exclude": ["dist", "node_modules", "**/*.spec.ts", "**/*.test.ts"],
   "compilerOptions": {
     "jsx": "react-jsx",
     "jsxImportSource": "react",
     "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "noImplicitOverride": true,
+    "exactOptionalPropertyTypes": true
   }
 }
